Migrate Dropdown API to DropdownLink, fixes various cosmetics issues like spacing and color (#4)
Created PackageSearchDropdownLink with a specific Style, injected top level

(cherry picked from commit 3adb17ede10a5fd141d4759fac4b35f50f4c42dd)
