+++ plugin/src/main/kotlin/com/jetbrains/packagesearch/plugin/ui/bridge/Components.kt
@@ -6,7 +6,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.TextLayoutResult
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontFamily
@@ -20,15 +19,12 @@ import com.intellij.icons.AllIcons
 import com.jetbrains.packagesearch.plugin.ui.PackageSearchMetrics
 import org.jetbrains.jewel.foundation.theme.JewelTheme
 import org.jetbrains.jewel.foundation.theme.LocalTextStyle
-import org.jetbrains.jewel.ui.component.Dropdown
+import org.jetbrains.jewel.ui.component.DropdownLink
 import org.jetbrains.jewel.ui.component.Icon
 import org.jetbrains.jewel.ui.component.IconButton
 import org.jetbrains.jewel.ui.component.MenuScope
 import org.jetbrains.jewel.ui.component.PopupMenu
 import org.jetbrains.jewel.ui.component.Text
-import org.jetbrains.jewel.ui.component.styling.DropdownColors
-import org.jetbrains.jewel.ui.component.styling.DropdownStyle
-import org.jetbrains.jewel.ui.theme.dropdownStyle
 
 @Composable
 fun LabelInfo(
@@ -71,7 +67,7 @@ fun LabelInfo(
 
 
 @Composable
-fun TextSelectionDropdown(
+fun PackageSearchDropdownLink(
     modifier: Modifier,
     menuModifier: Modifier,
     items: List<String>,
@@ -79,11 +75,11 @@ fun TextSelectionDropdown(
     enabled: Boolean,
     onSelection: (String) -> Unit,
 ) {
-    Dropdown(
+    DropdownLink(
         modifier = modifier,
         menuModifier = menuModifier.heightIn(max = PackageSearchMetrics.Dropdown.maxHeight),
         enabled = enabled && items.isNotEmpty(),
-        style = packageSearchDropdownStyle(),
+        style = LocalPackageSearchDropdownLinkStyle.current,
         menuContent = {
             items.forEach {
                 selectableItem(
@@ -94,17 +90,10 @@ fun TextSelectionDropdown(
                 }
             }
         },
-        content = {
-            Text(
-                modifier = Modifier.align(Alignment.CenterEnd),
-                text = content,
-                maxLines = 1,
-                overflow = TextOverflow.Clip,
-                textAlign = TextAlign.End
-            )
-        }
+        text = content
     )
 }
+
 
 @Composable
 internal fun PackageActionPopup(
@@ -140,37 +129,4 @@ internal fun PackageActionPopup(
             )
         }
     }
-}
-
-@Composable
-private fun packageSearchDropdownStyle(): DropdownStyle {
-    val currentStyle = JewelTheme.dropdownStyle
-    return DropdownStyle(
-        colors = DropdownColors(
-            background = Color.Transparent,
-            backgroundDisabled = Color.Transparent,
-            backgroundFocused = Color.Transparent,
-            backgroundPressed = Color.Transparent,
-            backgroundHovered = Color.Transparent,
-            content = currentStyle.colors.content,
-            contentDisabled = currentStyle.colors.contentDisabled,
-            contentFocused = currentStyle.colors.contentFocused,
-            contentPressed = currentStyle.colors.contentPressed,
-            contentHovered = currentStyle.colors.contentHovered,
-            border = Color.Transparent,
-            borderDisabled = Color.Transparent,
-            borderFocused = Color.Transparent,
-            borderPressed = Color.Transparent,
-            borderHovered = Color.Transparent,
-            iconTintDisabled = Color.Transparent,
-            iconTint = currentStyle.colors.iconTint,
-            iconTintFocused = currentStyle.colors.iconTintFocused,
-            iconTintPressed = currentStyle.colors.iconTintPressed,
-            iconTintHovered = currentStyle.colors.iconTintHovered,
-        ),
-        metrics = currentStyle.metrics,
-        icons = currentStyle.icons,
-        textStyle = currentStyle.textStyle,
-        menuStyle = currentStyle.menuStyle,
-    )
 }
